# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Road Trip Bingo Word Database API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /boards:
    get:
      operationId: listBoards
      tags: [boards]
      summary: List all bingo boards
      responses:
        "200":
          description: List of boards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardList"

    post:
      operationId: createBoard
      tags: [boards]
      summary: Create a new bingo board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBoardBody"
      responses:
        "201":
          description: Created board
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /snapshots:
    get:
      operationId: listSnapshots
      tags: [snapshots]
      summary: List all database snapshots
      responses:
        "200":
          description: List of snapshots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotList"

    post:
      operationId: createSnapshot
      tags: [snapshots]
      summary: Create a new database snapshot
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSnapshotBody"
      responses:
        "200":
          description: Created snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /snapshots/{id}/restore:
    post:
      operationId: restoreSnapshot
      tags: [snapshots]
      summary: Restore a database snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Snapshot restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestoreSnapshotResponse"
        "400":
          description: Invalid snapshot id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /snapshots/{id}:
    delete:
      operationId: deleteSnapshot
      tags: [snapshots]
      summary: Delete a database snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Snapshot deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                required:
                  - deleted
        "404":
          description: Snapshot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid snapshot id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/boards:
    get:
      operationId: listPublishedBoards
      tags: [public]
      summary: List published boards (read-only, versioned public API)
      description: >-
        Public, API-key-gated endpoint for external consumers (the iOS app).
        Returns only boards with published = true — draft/concept boards and
        unpublished edits are never visible here. Pass `since` (the highest
        contentVersion seen on a previous sync) to receive only boards whose
        content changed after that version.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: since
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          description: >-
            Only return boards with contentVersion greater than this value.
            Omit to list all published boards.
      responses:
        "200":
          description: Published boards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishedBoardList"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/boards/{id}/bundle:
    get:
      operationId: getBoardBundle
      tags: [public]
      summary: Get the self-contained content bundle for a published board
      description: >-
        Returns the full payload needed to play one board offline: board
        metadata plus its resolved word list (joined server-side, excluding
        soft-deleted words), a contentVersion, and a checksum over the word
        list. Only available for published boards.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Board bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardBundle"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Board not found or not published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Board:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        ageLevels:
          type: array
          items:
            type: string
          description: "e.g. Young, Kid, Tween"
        difficulty:
          type: ["string", "null"]
          description: "e.g. Easy, Medium, Hard"
        timeOfYear:
          type: ["string", "null"]
          description: "e.g. All Year, Summer, Winter"
        availability:
          type: ["string", "null"]
          description: "e.g. All, Seasonal"
        status:
          type: string
          enum: [active, draft, concept]
        published:
          type: boolean
          description: "Content-release gate: whether this board's content is included in the app-facing export. Independent of lifecycle status."
        publishedAt:
          type: ["string", "null"]
          format: date-time
          description: "When the board was last published, or null if unpublished"
        notes:
          type: ["string", "null"]
        wordCount:
          type: integer
          description: "Number of words assigned to this board"
        coverage:
          $ref: "#/components/schemas/BoardCoverage"
        preview:
          type: array
          items:
            $ref: "#/components/schemas/BoardPreviewWord"
          description: "Deterministic sample of words on this board (seeded by board id, stable across refetches)"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - published
        - publishedAt
        - createdAt
        - updatedAt
    BoardCoverage:
      type: object
      description: "Server-computed curation-quality indicator for a board's word pool"
      properties:
        status:
          type: string
          enum: [well-covered, needs-words, unbalanced]
        label:
          type: string
          description: "Human-readable summary, e.g. 'Needs more words (12/25)' or 'All Low findability'"
      required:
        - status
        - label
    BoardPreviewWord:
      type: object
      properties:
        id:
          type: integer
        word:
          type: string
        emoji:
          type: ["string", "null"]
      required:
        - id
        - word
        - emoji
    BoardList:
      type: object
      properties:
        boards:
          type: array
          items:
            $ref: "#/components/schemas/Board"
        total:
          type: integer
      required:
        - boards
        - total
    CreateBoardBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        ageLevels:
          type: array
          items:
            type: string
          default: []
        difficulty:
          type: ["string", "null"]
        timeOfYear:
          type: ["string", "null"]
        availability:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, draft, concept]
          default: active
        published:
          type: boolean
          default: false
        notes:
          type: ["string", "null"]
      required:
        - name
...
    Snapshot:
      type: object
      properties:
        id:
          type: string
          description: "Snapshot identifier, e.g. snap_1234567890"
        label:
          type: string
          description: "Human-readable label for the snapshot"
        createdAt:
          type: string
          format: date-time
        wordCount:
          type: integer
          description: "Number of words in the snapshot"
        sizeBytes:
          type: integer
          description: "File size of the SQL dump in bytes"
      required:
        - id
        - label
        - createdAt
        - wordCount
        - sizeBytes
    SnapshotList:
      type: object
      properties:
        snapshots:
          type: array
          items:
            $ref: "#/components/schemas/Snapshot"
      required:
        - snapshots
    CreateSnapshotBody:
      type: object
      properties:
        label:
          type: string
          description: "Human-readable label for the snapshot"
    SnapshotResponse:
      type: object
      properties:
        snapshot:
          $ref: "#/components/schemas/Snapshot"
      required:
        - snapshot
    RestoreSnapshotResponse:
      type: object
      properties:
        restored:
          type: boolean
        snapshot:
          $ref: "#/components/schemas/Snapshot"
      required:
        - restored
        - snapshot
...
    PublishedBoardSummary:
      type: object
      description: "Published-board metadata exposed on the public /v1 API"
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        ageLevels:
          type: array
          items:
            type: string
          description: "e.g. Young, Kid, Tween"
        difficulty:
          type: ["string", "null"]
          description: "e.g. Easy, Medium, Hard"
        timeOfYear:
          type: ["string", "null"]
          description: "e.g. All Year, Summer, Winter"
        contentVersion:
          type: integer
          description: "Globally-monotonic version stamped on the board's last content change"
        wordCount:
          type: integer
          description: "Number of non-deleted words linked to this board"
      required:
        - id
        - name
        - description
        - ageLevels
        - difficulty
        - timeOfYear
        - contentVersion
        - wordCount
    PublishedBoardList:
      type: object
      properties:
        boards:
          type: array
          items:
            $ref: "#/components/schemas/PublishedBoardSummary"
        total:
          type: integer
          description: "Number of boards in this response (after any `since` filter)"
        latestVersion:
          type: integer
          description: >-
            Highest contentVersion across ALL published boards (0 when none),
            regardless of the `since` filter. Persist this and pass it as
            `since` on the next sync.
      required:
        - boards
        - total
        - latestVersion
    BundleWord:
      type: object
      description: "A word entry inside a board bundle — gameplay content only, no editorial fields"
      properties:
        id:
          type: integer
        word:
          type: string
        spanish:
          type: ["string", "null"]
        emoji:
          type: ["string", "null"]
        age:
          type: ["string", "null"]
        findability:
          type: ["string", "null"]
        seasons:
          type: array
          items:
            type: string
        dayNight:
          type: array
          items:
            type: string
        regions:
          type: array
          items:
            type: string
        surroundings:
          type: array
          items:
            type: string
      required:
        - id
        - word
        - spanish
        - emoji
        - age
        - findability
        - seasons
        - dayNight
        - regions
        - surroundings
    BoardBundle:
      type: object
      description: "Self-contained payload for one published board: metadata + resolved word list"
      properties:
        board:
          $ref: "#/components/schemas/PublishedBoardSummary"
        words:
          type: array
          items:
            $ref: "#/components/schemas/BundleWord"
        contentVersion:
          type: integer
        checksum:
          type: string
          description: >-
            "sha256:<hex>" over the canonical JSON of the word list (words
            sorted by id, fixed key order). Lets the client verify a stored
            bundle or skip an unchanged re-download.
      required:
        - board
        - words
        - contentVersion
        - checksum
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Static API key for the public /v1 namespace. This key ships inside the
        iOS app binary — it is attribution/abuse control, not real access
        control. Admin routes use a separate password-gated bearer token.