# lib/api-zod/src/generated/types/boardBundle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Road Trip Bingo Word Database API
 * OpenAPI spec version: 0.1.0
 */
import type { BundleWord } from "./bundleWord";
import type { PublishedBoardSummary } from "./publishedBoardSummary";

/**
 * Self-contained payload for one published board: metadata + resolved word list
 */
export interface BoardBundle {
  board: PublishedBoardSummary;
  words: BundleWord[];
  contentVersion: number;
  /** "sha256:<hex>" over the canonical JSON of the word list (words sorted by id, fixed key order). Lets the client verify a stored bundle or skip an unchanged re-download. */
  checksum: string;
}