# lib/api-zod/src/generated/types/bundleWord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Road Trip Bingo Word Database API
 * OpenAPI spec version: 0.1.0
 */

/**
 * A word entry inside a board bundle — gameplay content only, no editorial fields
 */
export interface BundleWord {
  id: number;
  word: string;
  /** @nullable */
  spanish: string | null;
  /** @nullable */
  emoji: string | null;
  /** @nullable */
  age: string | null;
  /** @nullable */
  findability: string | null;
  seasons: string[];
  dayNight: string[];
  regions: string[];
  surroundings: string[];
}